feat(ThemeProvider): Embed fonts
