+++ package.json
@@ -37,7 +37,9 @@
     "polished": "^4.0.5",
     "react-i18next": "^11.8.12",
     "react-use": "^13.27.0",
-    "reakit": "^1.3.4"
+    "reakit": "^1.3.4",
+    "typeface-inconsolata": "^1.1.13",
+    "typeface-source-sans-pro": "^1.1.13"
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
