+++ src/components/ThemeProvider/ThemeContext/ThemeContext.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { DefaultTheme } from 'styled-components';
+
+export type ThemeContextProps = {
+	theme?: DefaultTheme;
+	switchTheme?: (theme: DefaultTheme) => void;
+};
+
+const ThemeContext = React.createContext<ThemeContextProps>({});
+
+export default ThemeContext;
