+++ src/components/ThemeProvider/ThemeContext/index.ts
@@ -1,0 +1,3 @@
+import ThemeContext from './ThemeContext';
+
+export default ThemeContext;
