+++ src/components/ThemeProvider/ThemeSwitcher/index.ts
@@ -1,0 +1,3 @@
+import ThemeSwitcher from './ThemeSwitcher';
+
+export default ThemeSwitcher;
