--- src/components/ThemeProvider/mixins.ts
@@ -1,7 +1,0 @@
-import { css } from 'styled-components';
-
-export const ellipsis = css`
-	text-overflow: ellipsis;
-	white-space: nowrap;
-	overflow: hidden;
-`;
